We have two test_cachefeed files. Merge them.
